task_04/pyproject.toml
[project]
name = "pdf-quiz-gen-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.6.1",
    "pydf>=12",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
